Fri, Oct 14, 2022, 12:25 AM +00:00

+++ chapter1/cash/cash.c
@@ -11,7 +11,7 @@
 int main(void)
 {
     // Asks user for the amount of change owed
-    float change;
+    int change;
     do
     {
         change = get_float("Change owed: ");
